bulk edit

+++ libgdx_sokoban/core/src/com/nopalsoft/sokoban/MainSokoban.kt
@@ -14,8 +14,8 @@ class MainSokoban : Game() {
     override fun create() {
         stage = Stage(
             StretchViewport(
-                Screens.SCREEN_WIDTH.toFloat(),
-                Screens.SCREEN_HEIGHT.toFloat()
+                Screens.SCREEN_WIDTH,
+                Screens.SCREEN_HEIGHT
             )
         )
         batcher = SpriteBatch()
